# agentscope-service/docs/service-api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentScope Service Invocation API",
    "version": "1.0.0"
  },
  "security": [
    {
      "endpointKey": []
    },
    {
      "platformBearer": []
    }
  ],
  "paths": {
...
    "/invoke/v1/invocations/{invocationId}/snapshot": {
      "get": {
        "summary": "Restore a consistent accumulated view",
        "operationId": "get_invocations_invocationId_snapshot",
        "parameters": [
          {
            "name": "invocationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Snapshot"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/invoke/v1/invocations/{invocationId}/actions": {
      "get": {
        "summary": "Read actions",
        "operationId": "get_invocations_invocationId_actions",
        "parameters": [
          {
            "name": "invocationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Enqueue durable actions command",
        "operationId": "post_invocations_invocationId_actions",
        "parameters": [
          {
            "name": "invocationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandInput"
              }
            }
          }
        }
      }
    },
    "/invoke/v1/invocations/{invocationId}/usage": {
      "get": {
        "summary": "Read usage",
        "operationId": "get_invocations_invocationId_usage",
        "parameters": [
          {
            "name": "invocationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/invoke/v1/invocations/{invocationId}/artifacts": {
      "get": {
        "summary": "Read artifacts",
        "operationId": "get_invocations_invocationId_artifacts",
        "parameters": [
          {
            "name": "invocationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agentId}": {
      "get": {
        "summary": "Read agent",
        "operationId": "get_api_v1_agents_agentId",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "platformBearer": []
          }
        ]
      },
      "patch": {
        "summary": "Update agent",
        "operationId": "patch_api_v1_agents_agentId",
        "parameters": [
          {
            "name": "agentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "platformBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "endpointKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "platformBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "registrationBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped Agent registration credential, not an Endpoint credential"
      }
    },
    "schemas": {
...
      "Invocation": {
        "type": "object",
        "required": [
          "id",
          "endpoint_id",
          "status",
          "mode"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "endpoint_id": {
            "type": "string",
            "format": "uuid"
          },
          "release_id": {
            "type": "string",
            "format": "uuid"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "accepted",
              "dispatching",
              "running",
              "waiting",
              "cancel_requested",
              "completed",
              "partial_succeeded",
              "failed",
              "cancelled",
              "timed_out"
            ]
          },
          "mode": {
            "enum": [
              "job",
              "conversation"
            ]
          },
          "result": {},
          "error": {
            "type": "object"
          }
        }
      },
      "Snapshot": {
        "type": "object",
        "required": [
          "as_of",
          "invocation",
          "items",
          "tools",
          "required_actions",
          "steps",
          "artifacts",
          "usage"
        ],
        "properties": {
          "as_of": {
            "type": "string"
          },
          "invocation": {
            "$ref": "#/components/schemas/Invocation"
          },
          "items": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "tools": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "required_actions": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "steps": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "artifacts": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          },
          "usage": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
...
      "CommandInput": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "expected_version": {
            "type": "integer"
          },
          "decision": {
            "enum": [
              "approved",
              "rejected"
            ]
          },
          "payload": {},
          "message": {
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "snapshot_url": {
            "type": "string"
          }
        }
      },
...
    }
  }
}